backend/requirements.txt
fastapi==0.115.6
uvicorn[standard]==0.34.0
sqlalchemy==2.0.36
psycopg2-binary==2.9.10
pydantic==2.10.6
pydantic-settings==2.7.0
python-dotenv==1.0.1
anthropic==0.42.0
alembic==1.14.0
python-multipart==0.0.20
email-validator==2.3.0
passlib[bcrypt]==1.7.4
bcrypt==4.2.1
python-jose[cryptography]==3.5.0
openai==2.14.0